refactor auth system for modularity and clarity
