+++ src/server/index.js
@@ -1,10 +1,13 @@
 var express  = require('express');
 var app = express();
 
+// Configure express
 require('./config/express.js')(app);
-require('./config/passport.js')(app);
+
+// Load routes onto app
 require('./routes')(app);
 
+// Listen to port
 app.listen(app.get('port'), function() {
   console.log('listening on port', app.get('port'));
 });
